// src/types/enums/constant-name.enum.ts
export enum ConstantNameEnum {
  INTERMEDIARY_TAX = "INTERMEDIARY_TAX",
  PAYMENT_TAX_PIX = "PAYMENT_TAX_PIX",
  MIN_ORDER_VALUE = "MIN_ORDER_VALUE",
  DELIVERY_HANDLING_DAYS = "DELIVERY_HANDLING_DAYS",
  PAYMENT_TIMEOUT_MINUTES = "PAYMENT_TIMEOUT_MINUTES",
  MAKER_ACCEPTANCE_DAYS = "MAKER_ACCEPTANCE_DAYS",
  REFUND_DELAY_TOLERANCE_DAYS = "REFUND_DELAY_TOLERANCE_DAYS",
  CONFIRMATION_TIMEOUT_DAYS = "CONFIRMATION_TIMEOUT_DAYS",
  NEW_DEADLINE_TIMEOUT_DAYS = "NEW_DEADLINE_TIMEOUT_DAYS",
  ADMIN_NOTIFICATION_EMAIL = "ADMIN_NOTIFICATION_EMAIL",
}